+++ packages/forms-transformer/test/lib/transforms/fix-choice.test.js
@@ -33,4 +33,3 @@ test('options should transform to the right format', (t) => {
 
   t.deepEqual(result, expected)
 })
-