+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+### Added
+- FT-141 | Store templates by form name in template service
+
 ## 1.0.0-alpha
 
 ### Added